Refactor OutputChannel and UIController to use uint64_t for frequency handling

+++ src/OutputChannel.h
@@ -14,9 +14,9 @@ public:
 
     void turnOff();
     void turnOn();
-    uint32_t setFrequency(uint32_t frequencyCentiHz);
-    uint32_t getActualFrequency() const;
-    uint32_t getSetFrequency() const;
+    uint64_t setFrequency(uint64_t frequencyCentiHz);
+    uint64_t getActualFrequency() const;
+    uint64_t getSetFrequency() const;
 
 private:
     const uint8_t selectPin;
@@ -24,13 +24,13 @@ private:
     TCB_t& tcb;
     Si5351& si5351;
     const si5351_clock siClock;
-    uint32_t actualFrequencyCentiHz_    = 0;
-    uint32_t setFrequencyCentiHz_       = 0;
-    const uint32_t SWITCHOVER_FREQUENCY = 400000UL;;
+    uint64_t actualFrequencyCentiHz_    = 0;
+    uint64_t setFrequencyCentiHz_       = 0;
+    const uint64_t SWITCHOVER_FREQUENCY = 400000UL;;
     boolean isOn_                       = false;
 
-    uint32_t setTCBFrequency(uint32_t frequencyCentiHz) const;
-    uint32_t setSiFrequency(uint32_t frequencyCentiHz) const;
+    uint64_t setTCBFrequency(uint64_t frequencyCentiHz) const;
+    uint64_t setSiFrequency(uint64_t frequencyCentiHz) const;
 };
